requirements.txt

flask
pyserial
